requirements.txt
pydantic
numpy
python-dotenv
openenv-core[core] >=0.2.2
fastapi >=0.115.0
uvicorn >=0.24.0
openai >=1.0.0

